@@ -15,17 +15,17 @@
     "last 2 Edge versions"
   ],
   "dependencies": {
-    "@angular/animations": "^8.2.14",
+    "@angular/animations": "^10.2.2",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/elements": "^8.2.14",
-    "@angular/forms": "^8.2.14",
+    "@angular/common": "^10.2.2",
+    "@angular/compiler": "^10.2.2",
+    "@angular/core": "^10.2.2",
+    "@angular/elements": "^10.2.2",
+    "@angular/forms": "^10.2.2",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/platform-browser": "^10.2.2",
+    "@angular/platform-browser-dynamic": "^10.2.2",
+    "@angular/router": "^10.2.2",
     "@types/diff2html": "0.0.4",
     "angular-feather": "^6.0.10",
     "angular2-uuid": "^1.1.1",
@@ -34,23 +34,23 @@
     "feather-icons": "^4.28.0",
     "jwt-decode": "^2.2.0",
     "monaco-editor": "^0.15.6",
-    "ngx-malihu-scrollbar": "^8.0.0",
-    "ngx-markdown": "^8.2.1",
+    "ngx-malihu-scrollbar": "^9.0.0",
+    "ngx-markdown": "~10.1.1",
     "path": "^0.12.7",
-    "rxjs": "^6.5.4",
+    "rxjs": "^6.6.3",
     "terser": "^3.17.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1002.0",
-    "@angular/cli": "^8.3.21",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@types/node": "^6.14.9",
-    "codelyzer": "^5.2.1",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.2",
+    "@angular/language-service": "^10.2.2",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "3.4.5"
+    "tslint": "~6.1.0",
+    "typescript": "4.0.5"
   }
 }
